MF-000 move calculateFareFromRouteSegment to FareService
